move waypointai derived functions to own file
